src/Ocelot/Middleware/Multiplexer/UserDefinedResponseAggregator.cs
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Ocelot.Configuration;

namespace Ocelot.Middleware.Multiplexer
{
    public class UserDefinedResponseAggregator : IResponseAggregator
    {
        private readonly IDefinedAggregatorProvider _provider;

        public UserDefinedResponseAggregator(IDefinedAggregatorProvider provider)
        {
            _provider = provider;
        }

        public async Task Aggregate(ReRoute reRoute, DownstreamContext originalContext, List<DownstreamContext> downstreamResponses)
        {
            var aggregator = _provider.Get(reRoute);

            if (!aggregator.IsError)
            {
                var aggregateResponse = await aggregator.Data
                    .Aggregate(downstreamResponses.Select(x => x.DownstreamResponse).ToList());

                originalContext.DownstreamResponse = aggregateResponse;
            }
            else
            {
                originalContext.Errors.AddRange(aggregator.Errors);
            }
        }
    }
}
